feat: enable serializeChunkAsync and deserializeChunkAsync
